% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCircleLegend.R
\name{addCircleLegend}
\alias{addCircleLegend}
\title{Add custom circle legend to leaflet map}
\usage{
addCircleLegend(
  map,
  title = "",
  values,
  scaling_fn,
  ...,
  color,
  weight,
  fillColor,
  fillOpacity,
  position = c("topright", "bottomright", "bottomleft", "topleft"),
  className = "info legend leaflet-control",
  layerId = NULL,
  group = NULL,
  data = leaflet::getMapData(map)
)
}
\arguments{
\item{map}{a leaflet map}

\item{title}{title of the legend}

\item{values}{the values used to generate circle radii}

\item{scaling_fn}{the scaling function used with \link[leaflet]{addCircleMarkers} to scale circle radii appropriately for leaflet}

\item{...}{additional arguements passed to scaling_fn}

\item{color}{stroke color}

\item{weight}{stroke width in pixels}

\item{fillColor}{fill color}

\item{fillOpacity}{fill opacity}

\item{position}{the position of the legend}

\item{className}{extra CSS class to append to the control, space separated}

\item{layerId}{the ID of the legend; subsequent calls to \code{addLegend}
or \code{addControl} with the same \code{layerId} will replace this
legend. The ID can also be used with \code{removeControl}.}

\item{group}{group name of a leaflet layer group}

\item{data}{a data object used for the leaflet map}
}
\value{
an object from \link[leaflet]{addControl}
}
\description{
To be used alongside \link[leaflet]{addCircleMarkers}
}
